refactor: videoStreamの取り回し、結果の描画関数の切り出し、videoとresultのcanvas部分をcomponent化
