+++ frontend/src/component/InferencePhase/Camera.js
@@ -2,21 +2,16 @@ import React, { useEffect, useRef, useState } from "react";
 import { useNavigate, useLocation } from 'react-router-dom';
 
 import useResultReceiver from "../../hooks/useResultReceiver";
-import useResultDrawer from "../../hooks/useResultDrawer";
+import ResultView from "./ResultView";
 
-function Camera({ streamRef, isStreamReady, videoSize }) {
-  const videoRef = useRef(null); // video要素への参照
-  const canvasRef = useRef(null); // canvas要素への参照
+function Camera({ streamRef, isStreamReady }) {
   const navigate = useNavigate();
   const location = useLocation();
   const label = location.state?.label;
 
-  const { drawResult, clearCanvas, canvasSize } = useResultDrawer(canvasRef, videoSize);
-  const { setupResultReceiver, deleteResultReceiver } = useResultReceiver(drawResult);
+  const { setupResultReceiver, deleteResultReceiver, setOnGetResult } = useResultReceiver();
 
   useEffect(() => {
-    videoRef.current.srcObject = streamRef.current;
-
     // ResultDrawerの初期化
     setupResultReceiver();
 
@@ -29,26 +24,7 @@ function Camera({ streamRef, isStreamReady, videoSize }) {
     <div style={{ padding: "20px" }}>
       <button onClick={() => navigate('/page2')}>ラベル選択に戻る</button>
       <h2>選択されたラベル: {label}</h2>
-
-      <div style={{ position: "relative", width: `${canvasSize.width}px`, height: `${canvasSize.height}px` }}>
-        {/* Video */}
-        <video
-          ref={videoRef}
-          autoPlay
-          playsInline
-          muted
-          width={canvasSize.width}
-          height={canvasSize.height}
-          style={{ position: "absolute", top: 0, left: 0 }}
-        />
-        {/* Canvas */}
-        <canvas
-          ref={canvasRef}
-          width={canvasSize.width}
-          height={canvasSize.height}
-          style={{ position: "absolute", top: 0, left: 0, pointerEvents: "none" }}
-        />
-      </div>
+      <ResultView isVideoStreamReady={isStreamReady} videoStreamRef={streamRef} setOnGetResult={setOnGetResult} />
     </div>
   );
 }
